+++ R/er_main.R
@@ -203,7 +203,7 @@ ecoregeling <- function(B_SOILTYPE_AGR, B_LU_BRP,B_LU_BBWP,
     out.farm[, s_er_water := pmin(dt.farm.thresholds$s_er_water_gold,s_er_water)]
     out.farm[, s_er_climate := pmin(dt.farm.thresholds$s_er_climate_gold,s_er_climate)]
     out.farm[, s_er_biodiversity := pmin(dt.farm.thresholds$s_er_biodiversity_gold,s_er_biodiversity)]
-    out.farm[, s_er_landscape := pmin(1,s_er_landscape)]
+    out.farm[, s_er_landscape := pmin(dt.farm.thresholds$s_er_landscape_gold,s_er_landscape)]
     out.farm[, s_er_farm_tot:= pmin(dt.farm.thresholds$s_er_farmtotal_gold,s_er_farm_tot)]
     
     # add thresholds