+++ tests/testthat/test-er.R
@@ -156,7 +156,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   test_that("check ecoregeling", {
     expect_equal(
       object = as.character(unlist(test$farm)),
-      expected = c(5.9,9.7,3.5,7.7,1,100,35,'silver',100,94,3.4,3.9,1.4,3.1,14,30,5.4,6.1,2.2,4.9,22,50,8.6,9.7,3.5,7.7,0,35,100),
+      expected = c(5.9,9.7,3.5,7.7,0,100,35,'silver',100,94,3.4,3.9,1.4,3.1,14,30,5.4,6.1,2.2,4.9,22,50,8.6,9.7,3.5,7.7,0,35,100),
       tolerance = 0.01)
   })